# Conflicts:
#	src/UI/Headquarters/WB.UI.Headquarters/API/Interviewer/InterviewerApiController.cs

+++ src/Core/SharedKernels/DataCollection/DataCollection/Commands/Interview/AnswerSingleOptionQuestionCommand.cs
@@ -5,9 +5,9 @@ namespace WB.Core.SharedKernels.DataCollection.Commands.Interview
 {
     public class AnswerSingleOptionQuestionCommand : AnswerQuestionCommand
     {
-        public decimal SelectedValue { get; private set; }
+        public int SelectedValue { get; private set; }
 
-        public AnswerSingleOptionQuestionCommand(Guid interviewId, Guid userId, Guid questionId, decimal[] rosterVector, DateTime answerTime, decimal selectedValue)
+        public AnswerSingleOptionQuestionCommand(Guid interviewId, Guid userId, Guid questionId, decimal[] rosterVector, DateTime answerTime, int selectedValue)
             : base(interviewId, userId, questionId, rosterVector, answerTime)
         {
             this.SelectedValue = selectedValue;
